Various build fixes

+++ src/core/bios.cpp
@@ -249,7 +249,7 @@ bool IsValidPSExeHeader(const PSEXEHeader& header, u32 file_size)
   if ((header.file_size + sizeof(PSEXEHeader)) > file_size)
   {
     Log_WarningPrintf("Incorrect file size in PS-EXE header: %u bytes should not be greater than %u bytes",
-                      header.file_size, file_size - sizeof(PSEXEHeader));
+                      header.file_size, static_cast<unsigned>(file_size - sizeof(PSEXEHeader)));
   }
 
   return true;
